Update molecular_class.py
I made some changes, mostly style (https://peps.python.org/pep-0008/). Let me know if you have any questions. Also keep in mind that it is not necessary to wrap all functions in a class. I would only use classes if it significantly improves the handling of parameters, and if the method actually uses the object in any way.